fix: polish handles check rpc
